requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
pandas
numpy
python-dotenv
pytest
pytest-asyncio